requirements.txt
aiohttp
async-timeout
attrs
cffi
chardet
discord.py
idna
lavalink
multidict
pycparser
PyNaCl
python-dotenv
six
typing-extensions
wavelink
yarl
unittest
json
apscheduler